+++ src/Zatomic.AI.Providers/Hyperbolic/HyperbolicChatRequest.cs
@@ -30,7 +30,7 @@ namespace Zatomic.AI.Providers.Hyperbolic
 		public float? RepetitionPenalty { get; set; }
 
 		[JsonProperty("seed", NullValueHandling = NullValueHandling.Ignore)]
-		public long? Seed { get; set; }
+		public decimal? Seed { get; set; }
 
 		[JsonProperty("stop", NullValueHandling = NullValueHandling.Ignore)]
 		public List<string> Stop { get; set; }
